requirements.txt
streamlit==1.54.0
httpx==0.28.1
python-dotenv==1.1.0
pandas==2.2.3
plotly==6.0.1
numpy==2.2.3
